feat: Moved sync account from menu to icon on transactions panel
