+++ packages/shared/components/modals/AccountActions.svelte
@@ -3,10 +3,9 @@
     import { openPopup } from 'shared/lib/popup'
     import { accountRoute } from 'shared/lib/router'
     import { AccountRoutes } from 'shared/lib/typings/routes'
-    import { WalletAccount, api, selectedAccountId, isSyncing } from 'shared/lib/wallet'
+    import type { WalletAccount } from 'shared/lib/wallet'
     import { getContext } from 'svelte'
     import type { Readable, Writable } from 'svelte/store'
-    import { NotificationData, NOTIFICATION_TIMEOUT_NEVER, showAppNotification, updateDisplayNotification } from 'shared/lib/notifications'
 
     const account = getContext<Readable<WalletAccount>>('selectedAccount')
     const accounts = getContext<Writable<WalletAccount[]>>('walletAccounts')
@@ -18,40 +17,7 @@
         accountRoute.set(AccountRoutes.Manage)
         isActive = false
     }
-    const handleSyncAccountClick = () => {
-        if (!$isSyncing) {
-            $isSyncing = true
 
-            const notificationData: NotificationData = {
-                type: 'info',
-                message: locale("general.accountSyncing"),
-                timeout: NOTIFICATION_TIMEOUT_NEVER
-            }
-
-            const notificationId = showAppNotification(notificationData)
-
-            api.syncAccount($selectedAccountId, {
-                onSuccess() {
-                    updateDisplayNotification(notificationId, {
-                        ...notificationData,
-                        message: locale("general.accountSyncComplete"),
-                        timeout: undefined
-                    })
-                    $isSyncing = false
-                },
-                onError(err) {
-                    updateDisplayNotification(notificationId, {
-                        ...notificationData,
-                        type: "error",
-                        message: locale(err.error),
-                        timeout: undefined
-                    })
-                    $isSyncing = false
-                },
-            })
-        }
-        isActive = false
-    }
     const handlViewAddressHistoryClick = () => {
         openPopup({ type: 'addressHistory', props: { account } })
         isActive = false
@@ -80,16 +46,8 @@
             <Icon icon="customize" classes="text-gray-500 ml-1 mr-3 group-hover:text-blue-500" />
             <Text smaller classes="group-hover:text-blue-500">{locale(`actions.customize_account`)}</Text>
         </button>
-        <!-- Sync -->
-        <button
-            on:click={() => handleSyncAccountClick()}
-            disabled={$isSyncing}
-            class={`group flex flex-row justify-start items-center py-3 px-3 w-full ${$isSyncing ? "cursor-auto opacity-30" : "hover:bg-blue-50"}`}>
-            <Icon icon="refresh" classes={`text-gray-500 ml-1 mr-3 ${$isSyncing ? "" : "group-hover:text-blue-500"}`} />
-            <Text smaller classes={`${$isSyncing ? "" : "group-hover:text-blue-500"}`}>{locale(`actions.sync_account`)}</Text>
-        </button>
         <!-- Address history -->
-       <!-- TODO: Implement and enable -->
+        <!-- TODO: Implement and enable -->
         <button
             disabled
             on:click={() => handlViewAddressHistoryClick()}
